on segment-click event pass the ID of current segment and anchestors
The annotation may reference not only leafs, but also ancestor
elements. E.g., when clicking on a span, an annotation may also be
asserted to the whole paragraph. Passing the ID of the span only,
results in not selecting the annotation to the paragraph. By passing
all IDs we find when walking to the root element, we make sure, that
the annotation to the paragraph is also noticed.