use eDisMax query parser
The DisMax query parser simplyfies searching over multiple fields like
needed for multilingual indexes, where we cannot use the
catch-all-field approach.

The eDisMax query parser combines the benefits of the
lucene (standard) query parser and the DisMax query parser, first of
all it makes "*" work on the q parameter, which is good for getting
all documents.